+++ Kokoro/API/Actions.cs
@@ -6,10 +6,21 @@ public partial interface IKokoroApi
 {
 	IActionApi Actions { get; }
 
-	public partial interface IActionApi
+	public interface IActionApi
 	{
 		CardAction MakeExhaustEntireHandImmediate();
 		CardAction MakePlaySpecificCardFromAnywhere(int cardId, bool showTheCardIfNotInHand = true);
 		CardAction MakePlayRandomCardsFromAnywhere(IEnumerable<int> cardIds, int amount = 1, bool showTheCardIfNotInHand = true);
+
+		List<CardAction> GetWrappedCardActions(CardAction action);
+		List<CardAction> GetWrappedCardActionsRecursively(CardAction action);
+
+		void RegisterWrappedActionHook(IWrappedActionHook hook, double priority);
+		void UnregisterWrappedActionHook(IWrappedActionHook hook);
 	}
+}
+
+public interface IWrappedActionHook
+{
+	List<CardAction>? GetWrappedCardActions(CardAction action);
 }